Updated the photos to be more optimized and moved around the home screen a bit
