+++ site/src/styles/global.scss
@@ -654,6 +654,11 @@ img {
   max-width: 100%;
 }
 
+picture {
+  display: block;
+  max-width: 100%;
+}
+
 h1,
 h2,
 h3,
@@ -750,6 +755,12 @@ h3 {
   min-height: 70svh;
 }
 
+.home-hero {
+  z-index: 2;
+  overflow: visible;
+  margin-bottom: clamp(6rem, 20svh, 14rem);
+}
+
 .hero-background,
 .hero-scrim {
   position: absolute;
@@ -759,6 +770,13 @@ h3 {
 }
 
 .hero-background {
+  display: block;
+}
+
+.hero-background img,
+.hero-background-image {
+  width: 100%;
+  height: 100%;
   object-fit: cover;
   object-position: center;
 }
@@ -782,9 +800,22 @@ h3 {
   }
 }
 
+.home-hero-content {
+  width: min(860px, 100%);
+  transform: translateY(clamp(9rem, 34svh, 20rem));
+}
+
+.brand-title-picture {
+  width: min(860px, 92vw);
+  max-width: 100%;
+  transform: translateX(-3%);
+}
+
 .brand-title-image {
-  width: min(520px, 92vw);
+  width: 100%;
+  height: auto;
   margin-bottom: 1.4rem;
+  object-fit: contain;
 }
 
 .product-logo {
@@ -1241,13 +1272,17 @@ h3 {
     min-height: 70svh;
   }
 
+  .home-hero-content {
+    transform: translateY(clamp(6.75rem, 25svh, 12rem));
+  }
+
   .product-hero {
     min-height: auto;
     padding: 2rem 1.25rem 1.5rem;
   }
 
   .brand-title-image {
-    width: min(360px, 100%);
+    width: 100%;
   }
 
   .app-card {
@@ -1501,11 +1536,17 @@ h3 {
   inset: 0;
   width: 100%;
   height: 100%;
-  object-fit: contain;
   filter: drop-shadow(0 28px 38px rgba(0, 0, 0, 0.42));
   opacity: 0.48;
   transform: translateX(-28%) rotate(-10deg) scale(0.88);
   transition: opacity 260ms ease, transform 320ms cubic-bezier(0.2, 0.8, 0.2, 1), filter 260ms ease;
+}
+
+.feature-scrub-image img,
+.feature-scrub-img {
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
 }
 
 .feature-scrub-image:nth-child(3) {
